// libs/store/Billing/actions-type.tsx
export enum BillingActionType {
    BILLING_ADDRESSES_SEARCHING = 'billing/BILLING_ADDRESSES_SEARCHING',
    BILLING_ADDRESSES_SEARCHING_SUCCESS = 'billing/BILLING_ADDRESSES_SEARCHING_SUCCESS',
    BILLING_ADDRESSES_SEARCHING_FAILED = 'billing/BILLING_ADDRESSES_SEARCHING_FAILED',
    ADD_BILLING_ADDRESSES = 'billing/ADD_BILLING_ADDRESSES',
    ADD_BILLING_ADDRESSES_SUCCESS = 'billing/ADD_BILLING_ADDRESSES_SUCCESS',
    ADD_BILLING_ADDRESSES_FAILED = 'billing/ADD_BILLING_ADDRESSES_FAILED ',
    DELETE_BILLING_ADDRESSES = 'billing/DELETE_BILLING_ADDRESSES',
    DELETE_BILLING_ADDRESSES_SUCCESS = 'billing/DELETE_BILLING_ADDRESSES_SUCCESS',
    DELETE_BILLING_ADDRESSES_FAILED = 'billing/DELETE_BILLING_ADDRESSES_FAILED '
}